# internal/api/handler/search.go
package handler

import (
	"context"
	"errors"
	"log"
	"slices"
	"time"

	"github.com/gofiber/fiber/v2"

	"github.com/strelov1/freehire/internal/dict/skillvec"
	"github.com/strelov1/freehire/internal/identity/auth"
	"github.com/strelov1/freehire/internal/identity/userprofile"
	"github.com/strelov1/freehire/internal/job/jobview"
	"github.com/strelov1/freehire/internal/platform/cache"
	"github.com/strelov1/freehire/internal/platform/db"
	"github.com/strelov1/freehire/internal/search/search"
	"github.com/strelov1/freehire/internal/search/suggest"
)

// searchHandlers serves the Meilisearch-backed job search surfaces: the keyword
// search, the agent search (always full descriptions in a selectable format), the
// similar-jobs read, and the facet distribution. search/facets are two narrow
// views of the same client, kept separate so the concerns stay decoupled; both
// nil when Meilisearch is unconfigured (the endpoints then report 503).
type searchHandlers struct {
	search searcher
	// descriptions loads full job descriptions by internal id, to rehydrate the
	// truncated search preview on the agent search endpoint. *db.Queries in
	// production; a fake in tests.
	descriptions jobDescriptions
	queries      *db.Queries
	facets       facetCounter
	// cache holds facet distributions for facetCacheTTL. Best-effort and optional:
	// a nil cache (search configured but no Redis) simply recomputes every request,
	// the shape every facet test but the caching ones runs under.
	cache cache.Cache
	// userProfile reads the caller's profile skills for the match sort. Nil disables
	// that sort entirely, which is exactly how it degrades everywhere else.
	userProfile profileReader
}

// profileReader loads a caller's profile. *userprofile.Service satisfies it; tests
// inject a fake. Only the skills are read here — the match sort has no opinion on
// location preferences or specializations.
type profileReader interface {
	Get(ctx context.Context, userID int64) (userprofile.Profile, error)
}

func newSearchHandlers(search searcher, facets facetCounter, queries *db.Queries, c cache.Cache, profiles profileReader) *searchHandlers {
	return &searchHandlers{
		search: search, descriptions: queries, queries: queries, facets: facets, cache: c,
		userProfile: profiles,
	}
}

// searchRecordTimeout bounds the fire-and-forget demand write. Short: it is one
// upsert on a small table, and a write still running past this is a database in
// trouble rather than a slow row.
const searchRecordTimeout = 5 * time.Second

func (h *searchHandlers) register(api fiber.Router, mw middleware) {
	// Literal routes are registered before the /jobs/:slug param route (see
	// Register) so they are not read as slugs.
	readLimit := publicReadLimiter(mw.throttler)
	api.Get("/jobs/search", readLimit, mw.optional, h.SearchJobs)
	// Agent search: same query as /jobs/search, but always full descriptions in a
	// selectable format for programmatic consumers. Public, like the other reads.
	api.Get("/agent/jobs/search", agentSearchLimiter(mw.throttler), h.AgentSearchJobs)
	api.Get("/jobs/facets", readLimit, h.JobFacets)
	api.Get("/jobs/:slug/similar", readLimit, h.SimilarJobs)
}

// searcher is the search backend the handler depends on. *search.Client
// satisfies it; tests inject a fake. A nil searcher means search is not
// configured (no MEILI_MASTER_KEY) and the endpoint reports 503. SimilarJobs and
// RecommendByVector were dropped from this interface when the Meili-backed
// jobs_semantic index they queried was removed — /similar now reads a precomputed
// pgvector lookup (see internal/api/handler/similar.go) and /me/recommendations was
// removed outright (see openspec/changes/drop-hybrid-search-pgvector-similar).
type searcher interface {
	Search(ctx context.Context, p search.SearchParams) (search.SearchResult, error)
}

// The pagination guard the search endpoints apply is maxPageWindow (handler.go), which
// they no longer own: it used to be a `maxSearchWindow` const here, and the Postgres-backed
// lists carried no equivalent at all — the asymmetry the 2026-09-14 deep-offset outage went
// through. It stays decoupled from the index's maxTotalHits (which now only sets how high
// the reported total may count), so the total can read the true filtered count while deep
// offset paging — the expensive part — stays refused.

// searchParams are the query params the search endpoints read themselves rather
// than hand to the filter: the query text, the sort directive and the pagination
// window. search.UnknownParams owns the filter vocabulary and nothing else, so
// each endpoint declares its own transport params beside itself (see
// facetsParams in facets.go, companiesParams in companies.go).
var searchParams = []string{"q", "sort", "order", "limit", "offset"}

// jobSearchParams is searchParams plus q_fields, the field-scoping parameter
// only the two job-search endpoints (not the swipe deck, which shares
// searchParams but never reads q_fields) understand.
var jobSearchParams = slices.Concat(searchParams, []string{"q_fields"})

// agentSearchParams is jobSearchParams plus the agent endpoint's
// response-format selector.
var agentSearchParams = slices.Concat(jobSearchParams, []string{"description_format"})

// ignoredParams reports the query params of this request that neither the filter
// nor the endpoint itself reads. They are echoed in the response meta instead of
// being refused: rejecting them would break saved searches and shared links that
// still carry retired params, while staying silent is what let a mistyped
// `country=it` pass for a search of Italy.
func ignoredParams(c *fiber.Ctx, own []string) []search.UnknownParam {
	return search.UnknownParams(queryValues(c), own)
}

// mergedJobSearchIgnored combines runJobSearch's own value-level ignored-param
// report — q_fields naming a field outside the searchable vocabulary, and/or any
// filter params the Meilisearch-rejection degrade dropped — with the ordinary
// name-level ignoredParams(c, own) report, into one sorted, capped list. The same
// merge both job-search endpoints need, named once so neither can apply it
// slightly differently.
func mergedJobSearchIgnored(c *fiber.Ctx, own []string, fromSearch []search.UnknownParam) []search.UnknownParam {
	return search.SortAndCap(append(fromSearch, ignoredParams(c, own)...))
}

// searchSortable is the allowlist of sort params mapped to their index attribute;
// anything else is ignored so a bad param cannot make Meilisearch reject the query.
var searchSortable = map[string]string{
	"created_at": "created_at",
	"posted_at":  "posted_at",
	// The "Most viewed" ordering. This entry must NEVER ship ahead of the live index
	// declaring view_count sortable (see search.facetSettings, and "Adding a sortable
	// attribute" in internal/search/search/AGENTS.md): Meilisearch rejects the entire
	// query for an undeclared sort attribute, and that package maps any Meili error to
	// a 500 — so the wrong deploy order breaks the page rather than ignoring the param.
	"view_count": "view_count",
	"salary_min": "enrichment.salary_min",
	"salary_max": "enrichment.salary_max",
}

// sortMatch is the profile-match sort. Unlike every value in searchSortable it names
// no index attribute: it ranks by cosine against the caller's own skill vector, so it
// is resolved from the request rather than looked up in that table.
const sortMatch = "match"

// matchVector builds the caller's skill vector for ?sort=match, or nil when the sort
// was not asked for or cannot be served.
//
// EVERY "cannot" degrades to the default feed: no session, no profile service wired,
// no profile, no skills, no recognised skills, no weights. A saved search or a shared
// link carrying sort=match must never 400 when opened by someone it cannot be served
// for — the same reason the jobs list ignores unknown filters rather than refusing
// them.
func (h *searchHandlers) matchVector(c *fiber.Ctx) []float32 {
	if c.Query("sort") != sortMatch || h.userProfile == nil {
		return nil
	}
	// auth.UserID is what mw.optional populates; it reports false for an anonymous
	// caller rather than erroring. requireUserID is the wrong helper here — its 401 is
	// exactly the outcome this must avoid.
	userID, ok := auth.UserID(c)
	if !ok {
		return nil
	}
	profile, err := h.userProfile.Get(c.Context(), userID)
	if err != nil || len(profile.Skills) == 0 {
		return nil
	}
	// ProfileVector, never JobVector: the ballast belongs to the stored side only.
	return skillvec.ProfileVector(profile.Skills)
}

// SearchJobs runs a full-text/faceted keyword search over the jobs index. It is
// public (unauthenticated) like the other job reads. Response: {"data": [job
// view...], "meta": {total, limit, offset}} — results carry public_slug and never
// the internal id.
func (h *searchHandlers) SearchJobs(c *fiber.Ctx) error {
	res, limit, offset, dropped, err := h.runJobSearch(c)
	if err != nil {
		return err
	}

	views := make([]jobview.Job, len(res.Hits))
	for i, hit := range res.Hits {
		views[i] = hit.Job
	}
	h.attachGhost(c, res.Hits, views)

	ignored := mergedJobSearchIgnored(c, jobSearchParams, dropped)
	return listResponseWithIgnored(c, views, res.Total, limit, offset, ignored)
}

// runJobSearch performs the request handling shared by both job-search endpoints:
// the availability check, the pagination-window guard, and the index query. It is
// the single place the query is built, so the public and agent search endpoints
// cannot drift. The availability and deep-pagination guards return a fiber *Error
// the caller can return directly; on success it returns the raw hits, the applied
// limit/offset, and every value-level ignored-param report gathered while building
// the query — q_fields naming a field outside the searchable vocabulary, plus any
// filter params dropped by the degrade below (nil in the ordinary case) — for the
// caller to merge into its own ignoredParams(c, own) report.
func (h *searchHandlers) runJobSearch(c *fiber.Ctx) (search.SearchResult, int, int, []search.UnknownParam, error) {
	if h.search == nil {
		return search.SearchResult{}, 0, 0, nil, fiber.NewError(fiber.StatusServiceUnavailable, "search is not available")
	}

	limit, offset, err := pageParamsWindowed(c, defaultLimit, maxLimit)
	if err != nil {
		return search.SearchResult{}, 0, 0, nil, err
	}

	vector := h.matchVector(c)
	sort := searchSort(c, vector != nil)
	filter := buildSearchFilter(c)
	qFields, qFieldsIgnored := search.QFieldsFromValues(queryValues(c))
	res, err := h.search.Search(c.Context(), search.SearchParams{
		Query:   c.Query("q"),
		Filter:  filter,
		Sort:    sort,
		Vector:  vector,
		QFields: qFields,
		Limit:   limit,
		Offset:  offset,
	})

	dropped := qFieldsIgnored
	if err != nil {
		// Degrade, not fail, when the filter itself is why Meilisearch refused the
		// query — the deploy window a filterable attribute is declared in code
		// before the live index catches up (AGENTS.md's "Adding a filterable
		// attribute") is the realistic trigger, not a malformed client value
		// (every facet/scalar value is already escaped before it reaches a filter
		// expression — see search.quote). filter == nil means there is nothing to
		// blame — retrying an unchanged query would just fail again identically.
		if filter == nil || !errors.Is(err, search.ErrBadQuery) {
			// RenderError renders a generic 500 (or 400 for ErrBadQuery); returning
			// the error keeps the failure cause visible to logging instead of
			// swallowing it.
			return search.SearchResult{}, 0, 0, nil, err
		}
		res, err = h.search.Search(c.Context(), search.SearchParams{
			Query: c.Query("q"), Sort: sort, Vector: vector, QFields: qFields, Limit: limit, Offset: offset,
		})
		if err != nil {
			return search.SearchResult{}, 0, 0, nil, err
		}
		// Which single param Meilisearch minded is not knowable from its own
		// error, so every active filter param is reported rather than a guess.
		dropped = append(dropped, search.ActiveFilterParams(queryValues(c))...)
	}

	h.recordQuery(c.Query("q"))
	return res, limit, offset, dropped, nil
}

// recordQuery notes what the catalogue was asked for, so the suggestion dictionary can
// rank by demand rather than only by supply.
//
// Three properties, all deliberate:
//
//   - It CANNOT fail the search. A write error is logged and discarded: the result is
//     what the visitor asked for and this is a by-product.
//   - It CANNOT delay the response. It runs on its own goroutine with its own bounded
//     context, because c.Context() is released when the handler returns.
//   - It identifies nobody. The row is the phrase and a counter; no user, no session,
//     no address — and `Recordable` refuses text that is not a search phrase at all,
//     so a pasted email address or job description never reaches the table.
//
// The key is the same normalisation the builder applies to mined titles, so a typed
// query and the title it names land on one row.
func (h *searchHandlers) recordQuery(raw string) {
	q := suggest.Title(raw)
	if !suggest.Recordable(q) || h.queries == nil {
		return
	}
	go func() {
		ctx, cancel := context.WithTimeout(context.Background(), searchRecordTimeout)
		defer cancel()
		if err := h.queries.RecordSearchQuery(ctx, q); err != nil {
			log.Printf("search: record query: %v", err)
		}
	}()
}

// searchSort builds the Meilisearch sort directive from ?sort=<field>&order=<dir>.
// Without a valid sort param, a no-text browse defaults to the freshest postings
// first (posted_at desc) — relevance is meaningless for an empty query — while a
// text query keeps relevance order (nil).
//
// ranked reports that the request carries a match vector. Meilisearch lets an explicit
// sort directive take precedence over vector ranking, so returning one here would
// silently discard the match order the caller asked for — not error, just quietly
// serve something else.
func searchSort(c *fiber.Ctx, ranked bool) []string {
	if ranked {
		return nil
	}
	attr, ok := searchSortable[c.Query("sort")]
	if !ok {
		if c.Query("q") == "" {
			return []string{"posted_at:desc"}
		}
		return nil
	}
	order := c.Query("order", "desc")
	if order != "asc" && order != "desc" {
		order = "desc"
	}
	return []string{attr + ":" + order}
}

// buildSearchFilter turns the request's facet query params into a Meilisearch
// filter by delegating to the shared, pure search.FilterFromValues — the same
// translation the notification matcher applies to a saved search's stored query,
// so the two cannot drift. Returns nil when no facet is set.
func buildSearchFilter(c *fiber.Ctx) any {
	return search.FilterFromValues(queryValues(c))
}

// attachGhost attaches the ghost signal to a page of search hits.
//
// Search is the surface where job cards are actually browsed, so leaving it out
// would mean the badge existed almost nowhere. It costs two extra reads per page:
// the outcome evidence (sparse — only jobs anyone reported or applied to) and the
// absence stamps.
//
// The stamps have to come from Postgres because they cannot be in the index:
// cmd/reindex pushes only documents whose content_hash moved, and no adapter writes
// ats_absent_at, so the column would never reach Meilisearch on its own. That is the
// same trap is_tech already fell into. The reality class, by contrast, IS on the
// document (search.FromJob promotes it), so it is read from the hit rather than
// recomputed.
//
// Best-effort: a failed lookup leaves the signal off the page rather than failing
// the search — but not silently. These stamps are a structural tier of this path,
// so a persistent failure here is the badge disappearing from every search result
// while the endpoint keeps answering 200; without the log there is nothing to see
// it by (see logGhostLookup for why cancellation is told apart on the context).
func (h *searchHandlers) attachGhost(c *fiber.Ctx, hits []search.JobDocument, views []jobview.Job) {
	if len(hits) == 0 || h.queries == nil {
		return
	}
	ids := make([]int64, len(hits))
	for i, hit := range hits {
		ids[i] = hit.ID
	}

	stampRows, err := h.queries.ListJobGhostStamps(c.Context(), ids)
	if err != nil {
		logGhostLookup(c.Context(), "absence stamps", len(ids), err)
		return
	}
	stamps := make(map[int64]db.ListJobGhostStampsRow, len(stampRows))
	for _, r := range stampRows {
		stamps[r.ID] = r
	}
	evidence := ghostEvidenceFor(c.Context(), h.queries, ids)

	now := time.Now()
	for i, hit := range hits {
		realityClass := ""
		if hit.Reality != nil {
			realityClass = hit.Reality.Class
		}
		row := stamps[hit.ID]
		views[i].Ghost = jobview.ClassifyGhost(jobview.GhostInput{
			Now: now,
			// Read from Postgres, not from the hit: a sweep-closed job stays in the
			// index until a reindex, whose timer is disabled, so the index cannot be
			// trusted to say a posting is still up. Without this a warning would
			// appear on postings already taken down.
			Closed:       row.ClosedAt.Valid,
			RealityClass: realityClass,
			ATSAbsentAt:  row.AtsAbsentAt.Time,
			HasATSAbsent: row.AtsAbsentAt.Valid,
			Evidence:     evidence[hit.ID],
		})
	}
}