# web/static/openapi.yaml
openapi: 3.0.3
info:
  title: freehire public API
  version: 1.1.0
  description: |
    Public, read-only freehire API: search IT jobs, read the live filter
    vocabulary, inspect job and company details, and find similar jobs. Every
    operation here is unauthenticated. The authenticated tracking surface
    (saved jobs, application stages, CV tooling) is deliberately excluded — this
    schema is the integration contract for search, and is also what the freehire
    custom GPT imports as an Action.

    ## Filter grammar

    Facet parameters are **repeatable**, and the same grammar applies to all of
    them:

    * `?skills=go&skills=kubernetes` — values of one facet are **ORed**: match
      either.
    * `?skills=go&skills=kubernetes&skills_mode=and` — `<facet>_mode=and` flips
      that facet to **AND**: match both. Not available for geography (see below).
    * `?source_exclude=adzuna` — `<facet>_exclude` removes matches. **Every**
      string facet in this schema supports its `_exclude` twin, whether or not it
      is listed below; only the most commonly used ones are declared explicitly.
    * Different facets **AND** together: `?work_mode=remote&seniority=senior` is
      remote *and* senior.

    ## Geography is one OR-group

    `regions`, `countries` and `cities` name a single concept — *where* — so
    their values join into **one OR-group** instead of intersecting.
    `?countries=gb&cities=London` therefore returns *United Kingdom **or**
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security: []
...
paths:
  /jobs/search:
    get:
      operationId: searchJobs
      tags:
        - Jobs
      summary: Search open jobs
      description: |
        Full-text and faceted search over open postings. Descriptions are the
        search index's truncated preview. Use `agentSearchJobs` when you need the
        full verbatim body of every hit, or `getJob` for the one posting the user
        picked.

        Closed postings are never in the index, so they never appear here.
      x-openai-isConsequential: false
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/QFields"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Order"
        - $ref: "#/components/parameters/Regions"
        - $ref: "#/components/parameters/Countries"
        - $ref: "#/components/parameters/Cities"
        - $ref: "#/components/parameters/WorkMode"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Seniority"
        - $ref: "#/components/parameters/Skills"
        - $ref: "#/components/parameters/SkillsMode"
        - $ref: "#/components/parameters/IsTech"
        - $ref: "#/components/parameters/RequiresClearance"
        - $ref: "#/components/parameters/AutoApplyAvailable"
        - $ref: "#/components/parameters/AIArchetype"
        - $ref: "#/components/parameters/RoleType"
        - $ref: "#/components/parameters/Collections"
        - $ref: "#/components/parameters/Reality"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/CompanySlug"
        - $ref: "#/components/parameters/EmploymentType"
        - $ref: "#/components/parameters/Relocation"
        - $ref: "#/components/parameters/EnglishLevel"
        - $ref: "#/components/parameters/EducationLevel"
        - $ref: "#/components/parameters/PostingLanguage"
        - $ref: "#/components/parameters/Domains"
        - $ref: "#/components/parameters/CompanyType"
        - $ref: "#/components/parameters/CompanySize"
        - $ref: "#/components/parameters/SalaryCurrency"
        - $ref: "#/components/parameters/SalaryPeriod"
        - $ref: "#/components/parameters/VisaSponsorship"
        - $ref: "#/components/parameters/SalaryMin"
        - $ref: "#/components/parameters/SalaryMax"
        - $ref: "#/components/parameters/ExperienceYearsMin"
        - $ref: "#/components/parameters/ExperienceYearsMax"
        - $ref: "#/components/parameters/PostedWithinDays"
        - $ref: "#/components/parameters/OpenWithinDays"
        - $ref: "#/components/parameters/RegionsExclude"
        - $ref: "#/components/parameters/CountriesExclude"
        - $ref: "#/components/parameters/WorkModeExclude"
        - $ref: "#/components/parameters/RoleTypeExclude"
        - $ref: "#/components/parameters/SkillsExclude"
        - $ref: "#/components/parameters/SourceExclude"
        - $ref: "#/components/parameters/CompanySlugExclude"
      responses:
        "200":
          description: Matching jobs with pagination metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListEnvelope"
        default:
          $ref: "#/components/responses/Error"
  /agent/jobs/search:
    get:
      operationId: agentSearchJobs
      tags:
        - Jobs
      summary: Search open jobs for agents (full descriptions)
      description: |
        Identical query surface to `searchJobs`, but every result carries the
        job's **full** description, read verbatim from the database rather than
        the search index's preview, in the format selected by
        `description_format`.

        Responses are several times larger per hit than `searchJobs`, so prefer a
        small `limit`.
      x-openai-isConsequential: false
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/QFields"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Order"
        - $ref: "#/components/parameters/DescriptionFormat"
        - $ref: "#/components/parameters/Regions"
        - $ref: "#/components/parameters/Countries"
        - $ref: "#/components/parameters/Cities"
        - $ref: "#/components/parameters/WorkMode"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Seniority"
        - $ref: "#/components/parameters/Skills"
        - $ref: "#/components/parameters/SkillsMode"
        - $ref: "#/components/parameters/IsTech"
        - $ref: "#/components/parameters/RequiresClearance"
        - $ref: "#/components/parameters/AutoApplyAvailable"
        - $ref: "#/components/parameters/AIArchetype"
        - $ref: "#/components/parameters/RoleType"
        - $ref: "#/components/parameters/Collections"
        - $ref: "#/components/parameters/Reality"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/CompanySlug"
        - $ref: "#/components/parameters/EmploymentType"
        - $ref: "#/components/parameters/Relocation"
        - $ref: "#/components/parameters/EnglishLevel"
        - $ref: "#/components/parameters/EducationLevel"
        - $ref: "#/components/parameters/PostingLanguage"
        - $ref: "#/components/parameters/Domains"
        - $ref: "#/components/parameters/CompanyType"
        - $ref: "#/components/parameters/CompanySize"
        - $ref: "#/components/parameters/SalaryCurrency"
        - $ref: "#/components/parameters/SalaryPeriod"
        - $ref: "#/components/parameters/VisaSponsorship"
        - $ref: "#/components/parameters/SalaryMin"
        - $ref: "#/components/parameters/SalaryMax"
        - $ref: "#/components/parameters/ExperienceYearsMin"
        - $ref: "#/components/parameters/ExperienceYearsMax"
        - $ref: "#/components/parameters/PostedWithinDays"
        - $ref: "#/components/parameters/OpenWithinDays"
        - $ref: "#/components/parameters/RegionsExclude"
        - $ref: "#/components/parameters/CountriesExclude"
        - $ref: "#/components/parameters/WorkModeExclude"
        - $ref: "#/components/parameters/RoleTypeExclude"
        - $ref: "#/components/parameters/SkillsExclude"
        - $ref: "#/components/parameters/SourceExclude"
        - $ref: "#/components/parameters/CompanySlugExclude"
      responses:
        "200":
          description: Matching jobs with pagination metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListEnvelope"
        default:
          $ref: "#/components/responses/Error"
  /jobs/facets:
    get:
      operationId: getJobFacets
      tags:
        - Jobs
      summary: Get live filter vocabulary and counts
      description: |
        Return facet values, counts, canonical skill slugs, and numeric ranges
        for the postings matching the same filter grammar the search endpoints
        take. Call this before applying any uncertain filter.

        A distribution is counted per facet and the wide-valued ones (`cities`,
        `skills`) dominate, so narrow with `facets=` whenever you read only a few.
        `company_slug` has no distribution — use `searchCompanies` for the company
        typeahead.
      x-openai-isConsequential: false
      parameters:
        - $ref: "#/components/parameters/Q"
        - name: facets
          in: query
          schema:
            type: string
          description: |
            Comma-separated facet names to count, e.g. `skills,seniority`. Omit
            for all of them. An unknown name is a 400. Cannot be combined with
            `disjunctive`.
        - name: disjunctive
          in: query
          schema:
            type: boolean
            default: false
          description: |
            Count each facet under the full filter MINUS its own selection, so a
            selected facet still shows its siblings' counts — the behaviour a
            live filter sidebar wants. Costs one query per facet.
        - $ref: "#/components/parameters/Regions"
        - $ref: "#/components/parameters/Countries"
        - $ref: "#/components/parameters/Cities"
        - $ref: "#/components/parameters/WorkMode"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Seniority"
        - $ref: "#/components/parameters/Skills"
        - $ref: "#/components/parameters/SkillsMode"
        - $ref: "#/components/parameters/IsTech"
        - $ref: "#/components/parameters/RequiresClearance"
        - $ref: "#/components/parameters/AutoApplyAvailable"
        - $ref: "#/components/parameters/AIArchetype"
        - $ref: "#/components/parameters/RoleType"
        - $ref: "#/components/parameters/Collections"
        - $ref: "#/components/parameters/Reality"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/CompanySlug"
        - $ref: "#/components/parameters/EmploymentType"
        - $ref: "#/components/parameters/Relocation"
        - $ref: "#/components/parameters/EnglishLevel"
        - $ref: "#/components/parameters/EducationLevel"
        - $ref: "#/components/parameters/PostingLanguage"
        - $ref: "#/components/parameters/Domains"
        - $ref: "#/components/parameters/CompanyType"
        - $ref: "#/components/parameters/CompanySize"
        - $ref: "#/components/parameters/SalaryCurrency"
        - $ref: "#/components/parameters/SalaryPeriod"
        - $ref: "#/components/parameters/VisaSponsorship"
        - $ref: "#/components/parameters/SalaryMin"
        - $ref: "#/components/parameters/SalaryMax"
        - $ref: "#/components/parameters/ExperienceYearsMin"
        - $ref: "#/components/parameters/ExperienceYearsMax"
        - $ref: "#/components/parameters/PostedWithinDays"
        - $ref: "#/components/parameters/OpenWithinDays"
      responses:
        "200":
          description: Facet distribution for matching jobs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacetsEnvelope"
        default:
          $ref: "#/components/responses/Error"
...
  /jobs/{slug}/similar:
    get:
      operationId: getSimilarJobs
      tags:
        - Jobs
      summary: Find jobs similar to a job
      description: |
        Nearest neighbours from a precomputed semantic list, nearest first.
        The list is built offline, so a neighbour that has since closed is
        dropped silently — a response may hold fewer items than `limit`.
      x-openai-isConsequential: false
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Job public_slug.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 6
          description: Maximum similar jobs to return. Values above the maximum are clamped.
      responses:
        "200":
          description: Similar jobs. This response carries no `meta` block.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Job"
        default:
          $ref: "#/components/responses/Error"
...
components:
...
  parameters:
    Q:
      name: q
      in: query
      schema:
        type: string
      description: |
        Full-text query over `title`, `company`, `description`, and `location`.

        Unquoted, multiple words match as an OR of (stemmed) tokens: any word
        matching in any of those fields is a hit. Double-quoted (e.g.
        `q="systems engineer"`), the words match as an AND, irrespective of
        their order or adjacency in the matched field — quoting does **not**
        request a contiguous phrase match. Use `q_fields` to narrow which
        fields `q` is matched against.

        Exact-phrase matching is not offered today, and the default relevance
        order (`sort` omitted) does not reliably rank a contiguous match above
        a scattered-token one within the same field — the index trades that
        precision for cheaper indexing over long-form descriptions. A caller
        that needs phrase precision must currently over-fetch and filter or
        rank client-side.
    QFields:
      name: q_fields
      in: query
      schema:
        type: string
      description: |
        Restrict `q` matching to a comma-separated subset of its four
        searchable fields: `title`, `company`, `description`, `location`. For
        example `q_fields=title` matches "systems" only in the job title, not
        in the company name or body — the only lever `q` has to exclude a
        posting like "STS Systems Defense" from a search for systems engineer
        roles.

        Naming any field outside that set (e.g. `q_fields=salary`) drops the
        parameter entirely — `q` runs unrestricted, across all four fields —
        and `q_fields` is reported in `meta.ignored_params`, the same
        convention every other unrecognized filter follows.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Page size. Values above the maximum are clamped, not rejected.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        # The window is on offset+limit, so this is the ceiling only for the smallest
        # possible limit. A generator cannot express the real bound; the description is
        # what states it.
        maximum: 10000
        default: 0
      description: |
        Rows to skip. On every list endpoint `offset + limit` may not exceed
        10000; deeper paging is a 400 ("pagination too deep"), never a clamped
        page — a 200 carrying rows you did not ask for would leave a client
        walking pages unable to tell it had reached the end. Narrow with
        filters rather than paging that far.
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum:
          - created_at
          - posted_at
          - view_count
          - salary_min
          - salary_max
      description: |
        Sort field. Omitted, a text query sorts by relevance and an empty query
        by `posted_at` descending. `view_count` orders by how many people have
        opened each posting.
    Order:
      name: order
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
      description: Sort direction. Ignored without a valid `sort`.
...
    Regions:
      name: regions
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - global
            - north_america
            - latam
            - eu
            - uk
            - mena
            - africa
            - apac
            - cis
            - none
      description: |
        Macro-regions. Country codes are NOT regions — use `countries`. `none`
        is the reserved value for postings with no resolved geography. Joins the
        geography OR-group with `countries` and `cities`.
    Countries:
      name: countries
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        ISO 3166-1 alpha-2, lowercase (`gb`, `de`). Matching is
        case-insensitive. Joins the geography OR-group.
    Cities:
      name: cities
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        Canonical city display names ("London", "Berlin") — resolve one with
        `searchCities` first. Joins the geography OR-group.
    WorkMode:
      name: work_mode
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - remote
            - hybrid
            - onsite
      description: |
        Work format, resolved from a deterministic dictionary. Absent on a
        posting whose format could not be resolved, so this filter narrows to
        postings that stated it.
    Category:
      name: category
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Role category slug from `getJobFacets`.
    Seniority:
      name: seniority
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - intern
            - junior
            - middle
            - senior
            - lead
            - staff
            - principal
            - c_level
...